Abstract routes to seperate routers in modules
